+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Mine and Slash random drops are now tier 1 minimum
+- Mine and Slash random drops should now be salvagable
+
 ## [1.3.8] - 2019-08-13
 ### Fixed
 - Mine and Slash compatibility race condition